adds potential fix to lambda error
